full api -1

+++ src/main/java/com/kf4b/bitlist/controller/TeamController.java
@@ -9,6 +9,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
+import java.util.Map;
+
 @RestController
 @RequestMapping("/teams")
 public class TeamController {
@@ -21,6 +24,18 @@ public class TeamController {
     @Autowired
     private JwtTokenUtil jwtTokenUtil;
 
+    private boolean checkPermissionFailure(User user, Team team, Team.Role role) {
+        if (user == null || team == null) {
+            return true;
+        }
+        Team.Role userRole = team.getMembers().get(user.getUserId());
+        if(role == Team.Role.ADMIN){
+            return userRole == null || !userRole.equals(Team.Role.ADMIN);
+        }else{
+            return userRole == null || (!userRole.equals(Team.Role.ADMIN) && !userRole.equals(Team.Role.MEMBER));
+        }
+    }
+
 
     /**
      * 创建团队
@@ -28,19 +43,35 @@ public class TeamController {
      * @return 新创建的团队对象
      */
     @PostMapping("")
-    public Team createTeam(@RequestBody Team team) {
+    public Team createTeam(@RequestHeader("Authorization") String token,
+                           @RequestBody Team team) {
+        // 检查token
+        User user = userService.getUserByHeader(token);
+        if (user == null) {
+            return null;
+        }
+        // 将当前用户作为管理员加入团队
+        Map<Integer, Team.Role> members = team.getMembers();
+        members.put(user.getUserId(), Team.Role.ADMIN);
+        // 保存团队
         teamService.updateTeamById(-1,team);
         return team;
     }
 
     /**
-     * 更新团队信息
-     * @param teamId 团队ID
+     * 获取团队信息
+     * @param userId 团队ID
      * @return 更新后的团队对象
      */
     @GetMapping("")
-    public Team getTeamById(@RequestParam Integer teamId) {
-        return teamService.getTeamById(teamId);
+    public List<Team> getTeamById(@RequestHeader("Authorization") String token,
+                            @RequestParam Integer userId) {
+        // 检查token
+        User user = userService.getUserByHeader(token);
+        if (user == null || !user.getUserId().equals(userId)) {
+            return null;
+        }
+        return teamService.getTeamsByUserId(userId);
     }
 
     /**
@@ -49,8 +80,17 @@ public class TeamController {
      * @param team 要更新的团队对象
      * @return 更新后的团队对象
      */
-    @PutMapping("{teamId}")
-    public Team updateTeam(@PathVariable Integer teamId, @RequestBody Team team) {
+    @PutMapping("/{teamId}")
+    public Team updateTeam(@RequestHeader("Authorization") String token,
+                           @PathVariable Integer teamId,
+                           @RequestBody Team team) {
+        // 检查token
+        User user = userService.getUserByHeader(token);
+        Team teamToUpdate = teamService.getTeamById(teamId);
+        if(checkPermissionFailure(user, teamToUpdate, Team.Role.ADMIN)){
+            return null;
+        }
+        // 更新
         team.setId(teamId);
         teamService.updateTeamById(teamId, team);
         return team;
@@ -61,7 +101,15 @@ public class TeamController {
      * @param teamId 团队ID
      */
     @DeleteMapping("{teamId}")
-    public boolean deleteTeam(@PathVariable Integer teamId) {
+    public boolean deleteTeam(@RequestHeader("Authorization") String token,
+                              @PathVariable Integer teamId) {
+        // 检查token
+        User user = userService.getUserByHeader(token);
+        Team team = teamService.getTeamById(teamId);
+        if(checkPermissionFailure(user, team, Team.Role.ADMIN)){
+            return false;
+        }
+        // 删除
         teamService.deleteTeam(teamId);
         return true;
     }
@@ -77,15 +125,8 @@ public class TeamController {
                               @RequestHeader("Authorization") String token) {
         // 检查token
         User user = userService.getUserByHeader(token);
-        if (user == null || !user.getUserId().equals(userId)) {
-            return false;
-        }
         Team team = teamService.getTeamById(teamId);
-        if (team == null) {
-            return false;
-        }
-        // 检查用户是否是团队的成员
-        if (!team.getMembers().containsKey(user.getUserId())) {
+        if(checkPermissionFailure(user, team, Team.Role.MEMBER)){
             return false;
         }
         team.getTasks().add(taskId);
@@ -102,13 +143,12 @@ public class TeamController {
                                    @RequestHeader("Authorization") String token) {
         // 检查token
         User user = userService.getUserByHeader(token);
-        if (user == null || !user.getUserId().equals(userId)) {
-            return false;
-        }
         Team team = teamService.getTeamById(teamId);
-        if (team == null) {
+        // 检查用户是否是团队的成员
+        if (team.getMembers().containsKey(user.getUserId())) {
             return false;
         }
+
         team.getPendingJoinRequests().add(user.getUserId());
         teamService.updateTeamById(teamId, team);
         return true;
@@ -122,15 +162,9 @@ public class TeamController {
                                 @RequestHeader("Authorization") String token) {
         // 检查token
         User user = userService.getUserByHeader(token);
-        if (user == null) {
-            return false;
-        }
         Team team = teamService.getTeamById(teamId);
-        if (team == null) {
-            return false;
-        }
         // 检查用户是否是团队的管理员
-        if (!team.getMembers().get(user.getUserId()).equals(Team.Role.ADMIN)) {
+        if(checkPermissionFailure(user, team, Team.Role.ADMIN)){
             return false;
         }
         // 检查用户是否存在
@@ -151,21 +185,16 @@ public class TeamController {
                                     @RequestHeader("Authorization") String token) {
         // 检查token
         User user = userService.getUserByHeader(token);
-        if (user == null) {
-            return false;
-        }
         Team team = teamService.getTeamById(teamId);
-        if (team == null) {
-            return false;
-        }
         // 检查用户是否是团队的管理员
-        if (!team.getMembers().get(user.getUserId()).equals(Team.Role.ADMIN)) {
+        if(checkPermissionFailure(user, team, Team.Role.ADMIN)){
             return false;
         }
         // 检查用户是否存在
         if (userService.getUserById(userIdToUpdate) == null) {
             return false;
         }
+        // 添加/更新成员角色
         team.getMembers().put(userIdToUpdate, role);
         teamService.updateTeamById(teamId, team);
         return true;
